requirements.txt
ultralytics==8.0.124
pillow-avif-plugin==1.3.1
mitmproxy==9.0.1
